% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_clust}
\alias{MoE_clust}
\title{Mixtures of Experts: Model-Based Clustering with Covariates}
\usage{
MoE_clust(data, G = 1:9, modelNames = NULL, gating = NULL,
  expert = NULL, control = MoE_control())
}
\arguments{
\item{data}{A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.}

\item{G}{An integer vector specifying the numbers of mixture components (clusters) to fit. Defaults to \code{G=1:9}.}

\item{modelNames}{A vector of character strings indicating the models to be fitted in the EM phase of clustering. The default is:\cr
\tabular{ll}{for univariate data \tab \code{c("E", "V")}\cr
for multivariate data \eqn{N > D}{N > D} \tab \code{mclust.options("emModelNames")}\cr
for high-dimensional multivariate data \eqn{N \leq D}{N <= D} \tab \code{c("EII", "VII", "EEI", "EVI", "VEI", "VVI")}
}
\cr
For single-component models these options reduce to:\cr
\tabular{ll}{
for univariate data \tab \code{"E"}\cr
for multivariate data \eqn{N > D}{N > D} \tab \code{c("EII", "EEI", "EEE")}\cr
for high-dimensional multivariate data \eqn{N \leq D}{N <= D}  \tab \code{c("EII", "EEI")}
}}

\item{gating}{A formula for determining the model matrix for the multinomial logistic regression in the gating network when covariates enter the mixing proportions. This will be ignored where \code{G=1}.}

\item{expert}{A formula for determining the model matrix for the multivariate WLS in the expert network when covariates are included in the component densities.}

\item{control}{A list of control parameters for the EM algorithm. The defaults are set by a call to \code{\link{MoE_control}}. Note that the argument \code{equalPro} will be ignored if \code{gating} network covariates are included.}
}
\value{
A list with the following named entries for the 'best' model as determined by the \code{criterion} within \code{\link{MoE_control}}:\cr
\itemize{
\item{\strong{mus} - }{The means of each component.}
\item{\strong{sigmas} - }{The variance parameters of each component.}
\item{\strong{Z} - }{The final responsibility matrix of probabilities of component membership.}
\item{\strong{pis} - }{The mixing proportion parameters: either a vector or, if \code{gating} covariates were supplied, an matrix with an entry for each observation and component.}
\item{\strong{BICs} - }{BIC values for every visited model.}
\item{\strong{ICLs} - }{ICL values for every visited model.}
\item{\strong{iters} - }{Total number of EM iterations for every visited model.}
\item{\strong{classes} - }{The vector of cluster labels for the chosen model.}
\item{\strong{best.model} - }{A list of details pertaining to the chosen model, incl. the mclust model type, the number of components, as well as "\code{bic}" and "\code{icl}" values.}
\item{\strong{log.likes} - }{The vector of increasing log-likelihood values for every EM iteration under the chosen model.}
\item{\strong{gating} - }{The \code{\link[nnet]{multinom}} coefficients of the \code{gating} network if \code{gating} covariates were supplied.}
\item{\strong{expert} - }{The multivariate WLS coefficients of the \code{expert} network if \code{expert} covariates were supplied.}
}
}
\description{
Fits Mixture of Experts models with mclust-family covariance structures. In other words, performs model-based clustering via the EM algorithm where covariates are allowed to enter neither, either, or both the mixing proportions (gating network) and/or component densities (expert network).
}
\details{
The function effectively allows 4 different types of Mixture of Expert model (as well as the different models in the mclust family, for each): i) the standard finite Gaussian mixture, ii) covariates only in the gating network, iii) covariates only in the expert network, iv) the full Mixture of Experts model with covariates entering both the mixing proportions and component densities. Note that it having the same covariates in both networks is allowed.
}
\examples{
\dontrun{
data(ais)
hema <- ais[,3:7]

# Fit a standard finite mixture model
m1   <- MoE_clust(hema, G=3)

# Allow covariates to enter the mixing proportions
m2   <- MoE_clust(hema, G=3, gating= ~ ais$sex + ais$BMI)

# Allow covariates to enter the component densities
m3   <- MoE_clust(hema, G=3, expert= ~ ais$sex + ais$BMI)

# Allow covariates to enter both the gating & expert network
m4   <- MoE_clust(hema, G=3, gating= ~ ais$sex + ais$BMI,
                  expert= ~ ais$sex + ais$BMI)

# Extract the model with highest BIC
BICS <- c(m1=m1$best.model$bic, m2=m2$best.model$bic, m3=m3$best.model$bic, m4=m4$best.model$bic)
get(gsub("\\\\..*", "", names(BICS)[which.max(BICS)]))$best.model
}
}
\seealso{
\code{\link{MoE_control}}, \code{\link{MoE_crit}}, \code{\link{MoE_dens}}
}
