% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_news}
\alias{MoE_news}
\title{Show the NEWS file}
\usage{
MoE_news()
}
\value{
The \code{MoEClust} \code{NEWS} file, provided the session is interactive.
}
\description{
Show the \code{NEWS} file of the \code{MoEClust} package.
}
\examples{
MoE_news()
}
\keyword{utility}
