% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot.MoEClust}
\alias{plot.MoEClust}
\title{Plot MoEClust Results}
\usage{
\method{plot}{MoEClust}(x, what = c("gpairs", "gating", "criterion"), ...)
}
\arguments{
\item{x}{An object of class \code{"MoEClust"} generated by \code{\link{MoE_clust}}, or an object of class \code{"MoECompare"} generated by \code{\link{MoE_compare}}.}

\item{what}{The type of graph requested:
\itemize{
\item{\code{"gpairs"} - }{A generalised pairs plot. To further customise this plot, arguments to \code{\link{MoE_gpairs}} can be supplied.}
\item{\code{"gating"} - }{The gating network. To further customise this plot, arguments to \code{\link{MoE_plotGate}} can be supplied.}
\item{\code{"criterion"} - }{The model selection criteria. To further customise this plot, arguments to \code{\link{MoE_plotCrit}} and \code{\link[mclust]{plot.mclustBIC}} can be supplied.}
}
By default, all of the above graphs are produced.}

\item{...}{Optional arguments to be passed to \code{\link{MoE_gpairs}}, \code{\link{MoE_plotGate}}, \code{\link{MoE_plotCrit}} and \code{\link[mclust]{plot.mclustBIC}}.}
}
\value{
The visualisation according to \code{"what"} of the results of a fitted \code{MoEClust} model.
}
\description{
Plot results for fitted MoE_clust mixture models with gating &/or expert network covariates: generalised pairs plots, model selection criteria, and the gating network are all currently visualisable.
}
\details{
For more flexibility in plotting, use \code{\link{MoE_gpairs}}, \code{\link{MoE_plotGate}}, \code{\link{MoE_plotCrit}} and \code{\link[mclust]{plot.mclustBIC}}
}
\note{
Other types of plots are available by first calling \code{\link{as.Mclust}} on the fitted object, and then calling \code{\link[mclust]{plot.Mclust}} on the results.
}
\examples{
data(ais)
sex <- ais$sex
BMI <- ais$BMI
res <- MoE_clust(ais[,3:7], gating= ~ sex, expert= ~ BMI, G=3, modelNames="VVE")

# Plot the gating network
plot(res, what="gating")

# Produce a generalised pairs plot
plot(res, what="gpairs")

# Modify the gpairs plot by passing arguments to MoE_gpairs()
plot(res, what="gpairs", response.type="uncertainty",
     scatter.type="lm", addEllipses=FALSE)
}
\seealso{
\code{\link{MoE_clust}}, \code{\link{MoE_gpairs}}, \code{\link{MoE_plotGate}}, \code{\link{MoE_plotCrit}}, \code{\link{as.Mclust}}, \code{\link[mclust]{plot.Mclust}}
}
\author{
Keefe Murphy - \href{keefe.murphy@ucd.ie}{<keefe.murphy@ucd.ie>}
}
