% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{quant_clust}
\alias{quant_clust}
\title{Quantile-Based Clustering for Univariate Data}
\usage{
quant_clust(x, G)
}
\arguments{
\item{x}{A vector of numeric data.}

\item{G}{The desired number of clusters.}
}
\value{
The vector of cluster labels.
}
\description{
Returns a quantile-based clustering for univariate data.
}
\examples{
data(CO2data)
quant_clust(CO2data$CO2, G=2)
}
\keyword{clustering}
