% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_clust}
\alias{MoE_clust}
\title{Mixtures of Experts: Model-Based Clustering with Covariates}
\usage{
MoE_clust(data, G = 1:9, modelNames = NULL, gating = NULL,
  expert = NULL, control = MoE_control(...), ...)
}
\arguments{
\item{data}{A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.}

\item{G}{An integer vector specifying the numbers of mixture components (clusters) to fit. Defaults to \code{G=1:9}.}

\item{modelNames}{A vector of character strings indicating the models to be fitted in the EM phase of clustering. The default is:\cr
\tabular{ll}{for univariate data \tab \code{c("E", "V")}\cr
for multivariate data \eqn{N > D}{N > D} \tab \code{mclust.options("emModelNames")}\cr
for high-dimensional multivariate data \eqn{N \leq D}{N <= D} \tab \code{c("EII", "VII", "EEI", "EVI", "VEI", "VVI")}
}
\cr
For single-component models these options reduce to:\cr
\tabular{ll}{
for univariate data \tab \code{"E"}\cr
for multivariate data \eqn{N > D}{N > D} \tab \code{c("EII", "EEI", "EEE")}\cr
for high-dimensional multivariate data \eqn{N \leq D}{N <= D}  \tab \code{c("EII", "EEI")}
}
The help file for \code{\link[mclust]{mclustModelNames}} further describes the available models (though the \code{"X"} in the single-component models will be coerced to \code{"E"} if supplied that way).}

\item{gating}{A formula for determining the model matrix for the multinomial logistic regression in the gating network when covariates enter the mixing proportions. This will be ignored where \code{G=1}. Interactions etc. are permitted. The specification of the LHS of the formula is ignored.}

\item{expert}{A formula for determining the model matrix for the multivariate WLS in the expert network when covariates are included in the component densities. Interactions etc. are permitted. The specification of the LHS of the formula is ignored.}

\item{control}{A list of control parameters for the EM and other aspects of the algorithm. The defaults are set by a call to \code{\link{MoE_control}}.}

\item{...}{An alternative means of passing control parameters via the named arguments of \code{\link{MoE_control}}. Do not pass the output from a call to \code{\link{MoE_control}} here!}
}
\value{
A list (of class \code{"MoEClust"}) with the following named entries, mostly corresponding to the chosen 'best' model (as determined by the \code{criterion} within \code{\link{MoE_control}}):\cr
\itemize{
\item{\code{mean} - }{The means of each component. Will be all \code{0} if there are \code{expert} network covariates.}
\item{\code{variance} - }{The variance parameters of each component.}
\item{\code{z} - }{The final responsibility matrix of probabilities of component membership.}
\item{\code{pro} - }{The mixing proportions: either a vector or, if \code{gating} covariates were supplied, a matrix with an entry for each observation (rows) and component (columns).}
\item{\code{classification} - }{The vector of cluster labels for the chosen model.}
\item{\code{BICs} - }{A matrix of BIC values with \code{length{G}} rows and \code{length(modelNames)} columns. May include missing entries: \code{NA} represents models which were not visited, \code{NaN} represents models which were terminated due to error.}
\item{\code{ICLs} - }{A matrix of ICL values with \code{length{G}} rows and \code{length(modelNames)} columns. May include missing entries: \code{NA} represents models which were not visited, \code{NaN} represents models which were terminated due to error.}
\item{\code{AICs} - }{A matrix of AIC values with \code{length{G}} rows and \code{length(modelNames)} columns. May include missing entries: \code{NA} represents models which were not visited, \code{NaN} represents models which were terminated due to error.}
\item{\code{iters} - }{A matrix giving the total number of EM iterations with \code{length{G}} rows and \code{length(modelNames)} columns. May include missing entries: \code{NA} represents models which were not visited, \code{NaN} represents models which were terminated due to error.}
\item{\code{best.model} - }{A list of details pertaining to the chosen model, incl. the \pkg{mclust} model type, the number of components (G), as well as "\code{bic}", "\code{icl}" and "\code{aic}" values (corresponding to the best model, which may not necessarily be the optimal for all three criteria). The supplied \code{gating} and \code{expert} formulas can also be found here, as well as the total degrees of freedom \code{df}.}
\item{\code{log.likes} - }{The vector of increasing log-likelihood values for every EM iteration under the chosen model.}
\item{\code{uncertainty} - }{The uncertainty associated with the classification.}
\item{\code{gating} - }{The \code{\link[nnet]{multinom}} regression coefficients of the \code{gating} network. If \code{gating} covariates were NOT supplied, this corresponds to a RHS of ~1, otherwise the supplied \code{gating} formula. The number of parameters to penalise by for \code{\link{MoE_crit}} is given by \code{gating$rank}.}
\item{\code{expert} - }{The multivariate WLS regression coefficients of the \code{expert} network. If \code{expert} covariates were NOT supplied, this corresponds to a RHS of ~1, otherwise the supplied \code{expert} formula. The number of parameters to penalise by for \code{\link{MoE_crit}} is given by \code{G * expert[[1]]$rank}.}
}
Dedicated \code{print} and \code{summary} functions exist for objects of class \code{"MoEClust"}.
}
\description{
Fits Mixture of Experts models with \pkg{mclust}-family covariance structures. In other words, performs model-based clustering via the EM algorithm where covariates are allowed to enter neither, either, or both the mixing proportions (gating network) and/or component densities (expert network).
}
\details{
The function effectively allows 4 different types of Mixture of Expert model (as well as the different models in the mclust family, for each): i) the standard finite Gaussian mixture, ii) covariates only in the gating network, iii) covariates only in the expert network, iv) the full Mixture of Experts model with covariates entering both the mixing proportions and component densities. Note that having the same covariates in both networks is allowed.\cr

While model selection in terms of choosing the optimal number of components and the \pkg{mclust} model type is performed within this function using one of the \code{criterion} options within \code{\link{MoE_control}}, choosing between multiple fits with different combinations of covariates can be done via \code{\link{MoE_compare}}.
}
\examples{
\dontrun{
data(ais)
hema  <- ais[,3:7]

# Fit a standard finite mixture model
m1    <- MoE_clust(hema, G=2:3)

# Allow covariates to enter the mixing proportions
m2    <- MoE_clust(hema, G=2:3, gating= ~ ais$sex + ais$BMI)

# Allow covariates to enter the component densities
m3    <- MoE_clust(hema, G=2:3, expert= ~ ais$sex)

# Allow covariates to enter both the gating & expert network
m4    <- MoE_clust(hema, G=2:3, gating= ~ ais$BMI, expert= ~ ais$sex)

# Extract the model with highest BIC
(comp <- MoE_compare(m1, m2, m3, m4))
(best <- get(comp$MoENames[1]))
summary(best)

# Visualise the results using the 'lattice' library
# require("lattice")
# z   <- factor(best$classification, labels=paste0("Cluster", seq_len(best$best.model$G)))
# splom(~ hema | ais$sex, groups=z)
# splom(~ hema | z, groups=ais$sex)}
}
\seealso{
\code{MoE_compare}, \code{\link{MoE_control}}, \code{\link{MoE_crit}}, \code{\link{MoE_estep}}, \code{\link{MoE_dens}}, \code{\link[mclust]{mclustModelNames}}
}
\author{
Keefe Murphy - \href{keefe.murphy@ucd.ie}{<keefe.murphy@ucd.ie>}
}
