% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_crit}
\alias{MoE_crit}
\title{MoEClust BIC and ICL Model-Selection Criteria}
\usage{
MoE_crit(modelName, loglik, N, D, G, gating.pen = G - 1, expert.pen = G * D,
  nn = NULL, delta = 0.5)
}
\arguments{
\item{modelName}{A character string indicating the model. The help file for \code{\link[mclust]{mclustModelNames}} describes the available models.}

\item{loglik}{The log-likelihood for a data set with respect to the Gaussian mixture model specified in the \code{modelName} argument.}

\item{N}{The number of observations in the data used to compute \code{loglik}.}

\item{D}{The dimension of the data used to compute \code{loglik}.}

\item{G}{The number of components in the Gaussian mixture model used to compute \code{loglik}.}

\item{gating.pen}{The number of parameters of the \emph{gating} network of the MoEClust model. Defaults to \code{G - 1}, which corresponds to no gating covariates. If covariates are included, this should be the number of regression coefficients in the fitted object. If there are no covariates and mixing proportions are further assumed to be present in equal proportion, \code{gating.pen} should be \code{0}.}

\item{expert.pen}{The number of parameters of the \emph{expert} network of the MoEClust model. Defaults to \code{G * D}, which corresponds to no expert covariates. If covariates are included, this should be the number of regression coefficients in the fitted object.}

\item{nn}{A vector of length \code{G} which sums to \code{N} giving the number of observations in each component. If supplied the ICL is also computed and returned, otherwise only the BIC.}

\item{delta}{Dirichlet hyperparameter for the prior on the mixing proportions. Defaults to 0.5. Only relevant for the ICL computation.}
}
\value{
A simplified array containing the BIC (and if \code{nn} is supplied, also the ICL) for each of the given input arguments.
}
\description{
Computes the BIC (Bayesian Information Criterion) and ICL (Integrated Complete Likelihood) for parameterized mixture of experts models given the log-likelihood, the dimension of the data, the number of mixture components in the model, the numbers of parameters in the gating and expert networks respectively, and, for the ICL, the numbers of observations in each component.
}
\details{
The function is vectorized with respect to the arguments \code{modelName} and \code{loglik}.\cr

If \code{model} is an object of class \code{"MoEClust"} with \code{G} components, the number of parameters for the \code{gating.pen} and \code{expert.pen} are \code{model$gating$rank} and \code{G * model$expert[[1]]$rank}, respectively.
}
\note{
In order to speed up repeated calls to the function inside \code{\link{MoE_clust}}, no checks take place.
}
\examples{
MoE_crit(modelName=c("VVI", "VVE", "VVV"),
         loglik=c(-4036.99, -3987.12, -3992.45),
         N=120, D=8, G=3, nn=c(54, 38, 28))

data(CO2data)
model <- MoE_clust(CO2data$CO2, G=1:2, expert= ~ CO2data$GNP)
G     <- model$best.model$G
(bic2 <- MoE_crit(modelName=model$best.model$modelName, loglik=max(model$log.likes),
                  N=length(CO2data$CO2), expert.pen=G * model$expert[[1]]$rank,
                  D=1, G=G, nn=tabulate(model$classes))["bic",])
identical(bic2, unname(model$best.model$bic))
}
\references{
Biernacki, C., Celeux, G., Govaert, G. (2000). Assessing a mixture model for clustering with the integrated completed likelihood. \emph{IEEE Trans. Pattern Analysis and Machine Intelligence}, 22(7): 719-725.
}
\seealso{
\code{\link{MoE_clust}}, \code{\link[mclust]{nVarParams}}, \code{\link[mclust]{mclustModelNames}}
}
\author{
Keefe Murphy - \href{keefe.murphy@ucd.ie}{<keefe.murphy@ucd.ie>}
}
