% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_dens}
\alias{MoE_dens}
\title{Density for Parameterised MVN Mixture of Experts Models}
\usage{
MoE_dens(modelName, data, mus, sigs, log.pis = 0, logarithm = TRUE)
}
\arguments{
\item{modelName}{A character string indicating the model. The help file for \code{\link[mclust]{mclustModelNames}} describes the available models.}

\item{data}{If there are no expert network covariates, \code{data} should be a numeric matrix or data frame, wherein rows correspond to observations and columns correspond to variables. If there are expert network covariates, this should be a list of length k containing matrices/data.frames of multivariate residuals for each component.}

\item{mus}{The mean for each component. If there is more than one component, this is a matrix whose k-th column is the mean of the k-th component of the mixture model. For the univariate models, this is a k-vector of means.}

\item{sigs}{A list of length k of variance parameters of the model. The components of this list depend on the specification of \code{modelName}.}

\item{log.pis}{If covariates enter the gating network, an \code{nrow(data)} times k matrix of mixing proportions, otherwise a k-vector of mixing proportions for the components of the mixture. \strong{Must} be on the log-scale in both cases. The default of \code{0} effectively means densities (or log-densities) aren't scaled by the mixing proportions.}

\item{logarithm}{A logical value indicating whether or not the logarithm of the component densities should be returned. This defaults to \code{TRUE}, otherwise component densities are returned, obtained from the component log densities by exponentiation.}
}
\value{
A list with two elements:
\itemize{
\item{\strong{dens} - }{A numeric matrix whose \code{[i,k]}-th entry is the density or log-density of observation \emph{i} in component \emph{k}, scaled by the mixing proportions.}
\item{\strong{norm} - }{A numeric vector giving the denominator in the density for each observation on the log-scale, computed efficiently via \code{\link[matrixStats]{rowLogSumExps}}, such that the sum of \code{norm} yields the log-likelihood.}
}
}
\description{
Computes densities of observations in parameterised MVN mixtures of experts.
}
\examples{
data(ais)
hema  <- ais[,3:7]
model <- MoE_clust(hema, G=2, gating= ~ ais$BMI, expert= ~ ais$sex)
Dens  <- MoE_dens(modelName=model$best.mod$modelName, data=hema,
                  mus=model$mus, sigs=model$sigmas, log.pis=log(model$pis))
}
\seealso{
\code{\link{MoE_clust}}, \code{\link[mclust]{mclustModelNames}}, \code{\link[matrixStats]{rowLogSumExps}}
}
