% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_estep}
\alias{MoE_estep}
\title{Compute the Z matrix and log-likelihood for MoEClust models}
\usage{
MoE_estep(Dens)
}
\arguments{
\item{Dens}{A numeric matrix whose \code{[i,k]}-th entry is the \strong{log}-density of observation \emph{i} in component \emph{k}, scaled by the mixing proportions, typically obtained by \code{\link{MoE_dens}} but this is not necessary.}
}
\value{
A list containing two elements:
\itemize{
\item{\strong{loglik} - }{The log-likelihood, computed efficiently via \code{\link[matrixStats]{rowLogSumExps}}}
\item{\strong{Z} - }{A matrix with N rows and G columns containing the probability of cluster membership for each of N observations and G clusters}
}
}
\description{
Takes the output from \code{\link{MoE_dens}} and computes the responsibility matrix Z, as well as the log-likelihood.
}
\examples{
data(ais)
hema  <- ais[,3:7]
model <- MoE_clust(hema, G=2:3, gating= ~ ais$BMI, model="VVE")
Dens  <- MoE_dens(modelName=model$best.mod$modelName, data=hema,
                  mus=model$mus, sigs=model$sigmas, log.pis=log(model$pis))

# Construct the Z matrix and compute the log-likelihood
Estep <- MoE_estep(Dens)
Z     <- Estep$Z
sum(Z - model$Z) == 0
Estep$loglik
}
\seealso{
\code{\link{MoE_dens}}, \code{\link{MoE_clust}}, \code{\link[matrixStats]{rowLogSumExps}}
}
