% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{MoE_gpairs}
\alias{MoE_gpairs}
\title{Generalised Pairs Plots for MoEClust Mixture Models}
\usage{
MoE_gpairs(res, response.type = c("points", "uncertainty", "density"),
  scatter.type = c("lm", "points"), conditional = c("stripplot", "boxplot"),
  addEllipses = TRUE, bg.col = c("ivory", "white", "snow", "snow",
  "cornsilk"), diagonal = c("on", "off"), outer.margins = list(bottom =
  unit(2, "lines"), left = unit(2, "lines"), top = unit(2, "lines"), right =
  unit(2, "lines")), outer.labels = NULL, outer.rot = c(0, 90),
  gap = 0.05, buffer = 0.02, scatter.pars = NULL, stripplot.pars = NULL,
  barcode.pars = NULL, mosaic.pars = NULL, axis.pars = NULL,
  diag.pars = NULL, ...)
}
\arguments{
\item{res}{An object of class \code{"MoEClust"} generated by \code{\link{MoE_clust}}, or an object of class \code{"MoECompare"} generated by \code{\link{MoE_compare}}.}

\item{response.type}{The type of plot desired for the scatter plots comparing continuous response variables. Defaults to \code{"points"}, but points can also be sized according to their associated clustering \code{"uncertainty"}, or the bivariate \code{"density"} can be displayed.}

\item{scatter.type}{A vector of length 2 giving the plot type for the upper and lower triangular portions of the plot, respectively, pertaining to the associated covariates. Defaults to \code{"lm"} for covariate vs. response panels and \code{"points"} otherwise. Only relevant for models with continuous covariates in the gating &/or expert network. \code{"ci"} and \code{"lm"} type plots are only produced for plots of covariates against response, and never response vs. response or covariate vs. covariate.}

\item{conditional}{A vector of length 2 giving the plot type for the upper and lower triangular portions of the plot, respectively, for plots involving a mix of categorical and continuous variables. Defaults to \code{"stripplot"} in the upper triangle and \code{"boxplot"} in the lower triangle (see \code{\link[lattice]{panel.stripplot}} and \code{\link[lattice]{panel.bwplot}}). Only relevant for models with categorical covariates in the gating &/or expert network. Comparisons of two categorical variables (which can only ever be covariates) are always displayed via mosaic plots (see \code{\link[vcd]{strucplot}}).}

\item{addEllipses}{Logical indicating whether to add ellipses with axes corresponding to the within-cluster covariances for the response data (defaults to \code{TRUE}).}

\item{bg.col}{A vector of length 5 containing background colours for plots against the MAP classification, response vs. response, covariate vs. response, response vs. covariate, and covariate vs. covariate panels, respectively. Defaults to \code{c("ivory", "white", "snow", "snow", "cornsilk")}.}

\item{diagonal}{By default, the diagonal to the top left to the bottom right is used for displaying the marginal distributions of variables. Specifying \code{"off"} will place the diagonal running from the top right down to the bottom left.}

\item{outer.margins}{A list of length 4 with units as components named bottom, left, top, and right, giving the outer margins; the defaults uses two lines of text. A vector of length 4 with units (ordered properly) will work, as will a vector of length 4 with numeric variables (interpreted as lines).}

\item{outer.labels}{The default is \code{NULL}, for alternating labels around the perimeter. If \code{"all"}, all labels are printed, and if \code{"none"}, no labels are printed.}

\item{outer.rot}{A 2-vector (\code{x}, \code{y}) rotating the top/bottom outer labels \code{x} degrees and the left/right outer labels \code{y} degrees. Only works for categorical labels of boxplot and mosaic panels.}

\item{gap}{The gap between the tiles; defaulting to 0.05 of the width of a tile.}

\item{buffer}{The fraction by which to expand the range of quantitative variables to provide plots that will not truncate plotting symbols. Defaults to 2 percent of range currently.}

\item{scatter.pars}{A list supplying select parameters for the continuous vs. continuous scatter plots. \code{NULL} is equivalent to \code{list(pch=res$classification, size=unit(0.25, "char"), col=res$classification, lcol=res$classification, border.col="black")}, where \code{lcol} gives the colour of the fitted lines when \code{scatter.type} is one of \code{"ci"} or \code{"lm"}. Note that \code{scatter.pars$size} will be overridden when \code{response.type} is \code{"uncertainty"}.}

\item{stripplot.pars}{\code{NULL} is equivalent to \code{list(pch=res$classification, size=unit(0.5, "char"), col=res$classification, jitter=TRUE)}.}

\item{barcode.pars}{\code{NULL} is equivalent to \code{list(nint=0, ptsize=unit(0.25, "char"), ptpch=1, bcspace=NULL, use.points=FALSE)}. See \code{\link[barcode]{barcode}}.}

\item{mosaic.pars}{\code{NULL}. Currently \code{shape, gp_labels, gp} and \code{gp_args} are passed through to \code{\link[vcd]{strucplot}} for producing mosaic tiles.}

\item{axis.pars}{\code{NULL} is equivalent to \code{list(n.ticks=5, fontsize=9)}.}

\item{diag.pars}{\code{NULL} is equivalent to \code{list(fontsize=9, show.hist=TRUE, hist.color=hist.color)}, where \code{hist.color} is a vector of length 5, giving the colours for the MAP classification, response variables, gating covariates, expert covariates, and covariates entering both networks, respectively. By default, response variables are \code{"black"}, covariates of any kind are \code{"grey"}, and the MAP classification is \code{"gold"}.}

\item{...}{Catches unused arguments.}
}
\value{
A generalised pairs plot.
}
\description{
Produces a matrix of plots showing pairwise relationships between continuous response variables and continuous/categorical associated covariates, as well as the clustering achieved, according to fitted MoEClust models.
}
\note{
For \code{MoEClust} models with more than one associated covariate (entering either network), fitted lines produced in continuous covariate vs. continuous response scatter plots via \code{scatter.type="lm"} or \code{scatter.type="ci"} will \strong{NOT} correspond to the coefficients in the expert network (\code{res$expert}).

\code{\link{plot.MoEClust}} is a wrapper to \code{\link{MoE_gpairs}} which accepts the default arguments, and also produces other types of plots.
}
\examples{
data(ais)
hema  <- ais[,3:7]
sex   <- ais$sex
BMI   <- ais$BMI

res   <- MoE_clust(hema, G=3, gating= ~ sex, expert= ~ sex + BMI, modelNames="EEE")
MoE_gpairs(res)

# Produce the same plot with a barchart in the lower triangle,
# with confidence intervals on the scatter plots,
# different colours for histograms of covariates in the gating/expert/both networks.
# Size points in the response vs. response panels by their clustering uncertainty.
# Also use different colours for response vs. covariate & covariate vs. response panels.
MoE_gpairs(res, scatter.type=c("ci", "points"), conditional=c("stripplot", "barcode"),
           diag.pars=list(hist.color=c("gold", "black", "cyan", "navy", "hotpink")),
           bg.col=c("oldlace", "white", "seashell", "whitesmoke", "cornsilk"))
}
\references{
Emerson, J.W., Green, W.A., Schloerke, B., Crowley, J., Cook, D., Hofmann, H. and Wickham, H. (2013). The Generalized Pairs Plot. \emph{Journal of Computational and Graphical Statistics}, 22(1):79-91.
}
\seealso{
\code{\link{MoE_clust}}, \code{\link{plot.MoEClust}}, \code{\link[lattice]{panel.stripplot}}, \code{\link[lattice]{panel.bwplot}}, \code{\link[vcd]{strucplot}}, \code{\link[barcode]{barcode}}
}
\author{
Keefe Murphy - \href{keefe.murphy@ucd.ie}{<keefe.murphy@ucd.ie>}
}
