% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_mahala}
\alias{MoE_mahala}
\title{Mahalanobis Distance Outlier Detection for Multivariate Response}
\usage{
MoE_mahala(fit, resids)
}
\arguments{
\item{fit}{A fitted \code{\link[stats]{lm}} model, inheriting either the \code{"mlm"} or \code{"lm"} class.}

\item{resids}{The residuals. Can be residuals for observations included in the model, or residuals arising from predictions on unseen data.}
}
\value{
A vector giving the Mahalanobis distance between fitted values and residuals for each observation.
}
\description{
Computes the Mahalanobis distance between the fitted values and residuals of linear regression models with multivariate or univariate responses.
}
\examples{
data(ais)
hema <- as.matrix(ais[,3:7])
mod  <- lm(hema ~ sex + BMI, data=ais)
res  <- hema - predict(mod)
MoE_mahala(mod, res)
}
\author{
Keefe Murphy - \href{keefe.murphy@ucd.ie}{<keefe.murphy@ucd.ie>}
}
