% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MoE_news}
\alias{MoE_news}
\title{Show the NEWS file}
\usage{
MoE_news(what = c("html", "md"))
}
\arguments{
\item{what}{A character string specifying whether the \code{"html"} (default) or Markdown \code{"md"} format is to be shown.}
}
\value{
The \code{MoEClust} \code{NEWS} file in the desired format.
}
\description{
Show the \code{NEWS} file of the \code{MoEClust} package.
}
\examples{
MoE_news()
}
\keyword{utility}
