% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{MoE_Uncertainty}
\alias{MoE_Uncertainty}
\title{Plot Clustering Uncertainties}
\usage{
MoE_Uncertainty(res, truth = NULL, ...)
}
\arguments{
\item{res}{An object of class \code{"MoEClust"} generated by \code{\link{MoE_clust}}, or an object of class \code{"MoECompare"} generated by \code{\link{MoE_compare}}.}

\item{truth}{An optional argument giving the true classification of the data. When \code{truth} is supplied, misclassified observations are highlighted in a different colour, otherwise observations with uncertainty greater than \code{1/res$G} are given in a different colour.}

\item{...}{Catches unused arguments.}
}
\value{
A plot showing the clustering uncertainty of each observation.
}
\description{
Plots the clustering uncertainty for every observation from a fitted \code{"MoEClust"} model.
}
\details{
The y-axis of this plot runs from \code{0} to \code{1 - 1/res$G}, with a horizontal line also drawn at \code{1/res$G}. Uncertainties greater than this value are given a different colour when \code{truth} is not supplied, otherwise misclassified observations are given a different colour.
}
\note{
\code{\link{plot.MoEClust}} is a wrapper to \code{\link{MoE_plotGate}} which accepts the default arguments, and also produces other types of plots.

An alternative means of visualising clustering uncertainties (at least for multivariate data) is provided by the functions \code{MoE_gpairs} and \code{plot.MoEClust}, specifically when their argument \code{response.type} is given as \code{"uncertainty"}.
}
\examples{
data(ais)
res <- MoE_clust(ais[,3:7], gating= ~ sex, G=3, modelNames="EEE", network.data=ais)
MoE_Uncertainty(res)

# Let's assume the true clusters correspond to sex
MoE_Uncertainty(res, truth=ais$sex)
}
\seealso{
\code{\link{MoE_clust}}, \code{\link{MoE_gpairs}}, \code{\link{plot.MoEClust}}
}
\author{
Keefe Murphy - <\email{keefe.murphy@ucd.ie}>
}
\keyword{plotting}
