% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{predict.MoEClust}
\alias{predict.MoEClust}
\alias{residuals.MoEClust}
\title{Predictions for MoEClust models}
\usage{
\method{predict}{MoEClust}(object,
        newdata,
        resid = FALSE,
        ...)

\method{residuals}{MoEClust}(object,
          newdata,
          ...)
}
\arguments{
\item{object}{An object of class \code{"MoEClust"} generated by \code{\link{MoE_clust}}, or an object of class \code{"MoECompare"} generated by \code{\link{MoE_compare}}. Predictions for models with a noise component are facilitated here too.}

\item{newdata}{A list with two \emph{named} components, each of which must be a \code{data.frame} or \code{matrix} with named columns, giving the data for which predicitions are desired.
\describe{
\item{\code{new.x}}{The new covariates for the \code{gating} &/or \code{expert} networks. \strong{Must} be supplied when \code{newdata$new.y} is supplied.}
\item{\code{new.y}}{(Optional) response data. When supplied, cluster and response prediction is based on both \code{newdata$new.x} and \code{newdata$new.y}, otherwise only on the covariates in \code{newdata$new.x}.}
}
If supplied as a list with elements \code{new.x} and \code{new.y}, both \strong{must} have the same number of rows.

Alternatively, a single \code{data.frame} or \code{matrix} can be supplied and an attempt will be made to extract & separate covariate and response columns (\emph{if any}) into \code{newdata$new.x} and \code{newdata$new.y} based on the variable names in \code{object$data} and \code{object$net.covs}.

When \code{newdata} is not supplied in any way, the covariates and response variables used in the fitting of the model are used here.}

\item{resid}{A logical indicating whether to return the residuals also. Defaults to \code{FALSE}. Only allowed when response variables are supplied in some form. The function \code{residuals} is a wrapper to \code{predict} with the argument \code{resid} set to \code{TRUE}, with only the residuals returned.}

\item{...}{Catches unused arguments.}
}
\value{
A list with the following named components, regardless of whether \code{newdata$new.x} and \code{newdata$new.y} were used, or \code{newdata$new.x} only.
\item{\code{y}}{Fitted values of the response variables.}
\item{\code{z}}{A matrix whose \code{[i,k]}-th entry is the probability that observation \emph{i} of the \code{newdata} belonds to the \emph{k}-th component..}
\item{\code{classification}}{The vector of predicted cluster labels for the \code{newdata}.}

When \code{residuals} is called, only the residuals are returned; when \code{predict} is called with \code{resid=TRUE}, the list above will also contain the element \code{resids}, containing the residuals.
}
\description{
Predicts both cluster membership probability and fitted response values from a \code{MoEClust} model, using covariates and response data, or covariates only. The MAP classification is also reported in both cases.
}
\note{
Predictions can also be made for models with noise components, in which case \code{z} will include the probability of belonging to \code{"Cluster0"} & \code{classification} will include labels with the value \code{0} for observations classified as noise (if any). Note that calculating \code{ystar} for models with a noise component involves renormalising the rows of \code{zstar} such that non-noise columns sum to 1, but the \emph{unnormalised} \code{zstar} is what's reported.
}
\examples{
data(ais)

# Fit a MoEClust model and predict the same data
res     <- MoE_clust(ais[,3:7], G=2, gating=~BMI, expert=~sex,
                     modelNames="EVE", network.data=ais)
pred1   <- predict(res)
pred1$classification

# Remove some rows of the data for prediction purposes
ind     <- sample(1:nrow(ais), 5)
dat     <- ais[-ind,]

# Fit another MoEClust model to the retained data
res2    <- MoE_clust(dat[,3:7], G=3, gating=~BMI + sex,
                     modelNames="EEE", network.data=dat)

# Predict held back data using the covariates & response variables
pred2   <- predict(res2, newdata=ais[ind,])
# pred2 <- predict(res2, newdata=list(new.y=ais[ind,3:7],
#                                     new.x=ais[ind,c("BMI", "sex")]))
pred2$y

# Get the residuals
residuals(res2, newdata=ais[ind,])

# Predict held back data using only the covariates
pred3   <- predict(res2, newdata=list(new.x=ais[ind,c("BMI", "sex")]))
# pred3 <- predict(res2, newdata=ais[ind,c("BMI", "sex")])
pred3$z
}
\references{
K. Murphy and T. B. Murphy (2017). Parsimonious Model-Based Clustering with Covariates. \emph{To appear}. <\href{https://arxiv.org/abs/1711.05632}{arXiv:1711.05632}>.
}
\seealso{
\code{\link{MoE_clust}}
}
\author{
Keefe Murphy - <\email{keefe.murphy@ucd.ie}>
}
\keyword{clustering}
\keyword{main}
\keyword{utility}
